+++ runtime/src/bank/builtin_programs.rs
@@ -66,7 +66,7 @@ mod tests {
 mod tests_core_bpf_migration {
     use {
         crate::bank::{
-            builtins::{core_bpf_migration::tests::TestContext, BUILTINS, STATELESS_BUILTINS},
+            builtins::core_bpf_migration::tests::TestContext,
             test_utils::goto_end_of_slot,
             tests::{create_genesis_config, new_bank_from_parent_with_bank_forks},
             Bank,
@@ -74,6 +74,7 @@ mod tests_core_bpf_migration {
         solana_builtins::{
             core_bpf_migration::CoreBpfMigrationConfig,
             prototype::{BuiltinPrototype, StatelessBuiltinPrototype},
+            BUILTINS, STATELESS_BUILTINS,
         },
         solana_feature_set::FeatureSet,
         solana_program_runtime::loaded_programs::ProgramCacheEntry,
